POSIX: Move assureDirectoryExists to posix-fs{.h,.cpp}.
